Computes long-term equilibrium yields for Rpath-package models

Viable fn to be used as input for nash when the Rpath-package is used as operating ecological model (see Lucey2020 for details) . fn_rpath takes the harvesting rates as the numeric type vector par returning simulated yields at equilibrium.

Usage

fn_rpath(
  par,
  simul.years = 100,
  aged.str = TRUE,
  data.years,
  IDnames,
  rsim.mod,
  rpath.params,
  avg.window = 10,
  integration.method = "RK4"
)

Arguments

par

Double type numeric vector of harvesting rates of length equal to the number of harvested species for which the NE is desired.

simul.years

Forward simulation time in years.

aged.str

Logical TRUE/FALSE if multistanza functional groups are included in the model.

data.years

Numeric vector indicating the years worth of data used to parameterise the Rpath model.

IDnames

Character vector with the names of the species for which Nash Equilibrium harvesting rates are computed. Note that these names must coincide with the ones used during the construction of the Rpath model.

rsim.mod

Rpath's rsim.scenario object.

rpath.params

Rpath's rpath.parameters object.

avg.window

Numeric type vector indicating the time window used to average equilibrium yields.

integration.method

Numerical integration routine used to solve the rsim.mod object. Character vector with values (i) `RK4` or (ii) `AB`.

Value

The function fn_rpath returns an atomic vector of real double-precision long-term yields.

Details

The avg.window argument becomes useful in case the dynamics of the model reaches a steady state (e.g. a limit cycle) rather than a stable point attractor.

The numerical integration methods implemented in the rsim.run are the 4th order Runge-Kutta (RK4) and the two-step Adams-Bashforth (AB) method. The trade-off between both methods is accuracy and speed, with RK4 being more accurate but slower than the AB method (see Lucey2020 for details) .

fn_rpath works for aged structure models in which multistanza species are partition into adults and juveniles. Then a fixed harvesting rate ratio between adults and juveniles is computed and retained when the function nash is called.

References

Lucey SM, Gaichas SK, Aydin KY (2020). “Conducting reproducible ecosystem modeling using the open source mass balance model Rpath.” Ecological Modelling, 427, 109057. doi:10.1016/j.ecolmodel.2020.109057 .