Computes harvesting rates at the Nash Equilibrium (NE)

Takes a numeric vector of initial harvesting rates to be optimised (`par`) by evaluating the objective function that is to be maximised (`fn`). Its structure in terms of arguments and outputs is similar to that of optim.

Usage

nash(
  par,
  fn,
  ...,
  method = "LV",
  yield.curves = FALSE,
  conv.criterion = 0.001,
  Bcons = 0,
  F.increase = 0.1,
  progress = TRUE,
  track = FALSE
)

Arguments

par

Double type numeric vector of harvesting rates of length equal to the number of harvested species for which the NE is desired.

fn

Function that runs the multispecies/ecosystem model with par as input and returns simulated yields at equilibrium.

...

Further arguments to be passed to fn.

method

Method utilised to compute the Nash Equilibrium: (i) `LV` or (ii) `round-robin` method (see `Details` section for specifics).

yield.curves

Logical TRUE/FALSE if equilibrium yield curves for each of the optimised species are to be computed.

conv.criterion

Absolute convergence tolerance set by default to \(< 0.001\).

Bcons

Constraints for biodiversity conservation. Double type numeric vector set to \(0\) by default.

F.increase

Double type numeric vector indicating the step size used to compute the effective interaction matrix \(M\).

progress

Logical that if TRUE information on the progress of the optimisation is produced.

track

Logical that if TRUE will return all fishing mortality rates that where iteratively computed during the search.

Value

The function nash returns a list with the following components:

par

Harvesting rates at the NE.

value

Yield values of fn corresponding to the optimised par.

counts

Number of fn evaluations until NE.

convergence

Statement indicating the number of iterations for conv.criterion to be reached.

Details

For ecosystem models where there is some interest in keeping some or all harvested species above a certain biomass state limit, Bcons should be populated with non-zero biomass values. The length of this vector must be the same as par and set to non-zero where relevant. In the literature, is common practice to fixed such constraints of biodiversity conservation (Matsuda and Abrams 2006) as a fraction (e.g. \(0.1-0.25\)) of the unfished biomass \(B_0\); biomass threshold at which a stock is considered collapsed (Worm et al. 2009) .

Equilibrium yield curves are obtained for each \(i\) species by applying different harvesting values to \(i\) whilst keeping the other species \(j\) at the optimised par levels (i.e. at \(\mathbf{F_{Nash}}\)). The harvesting values applied to \(i\) run from \(0\) to \(F_{Nash,i}\times 2\) with a desired sequence length of length.out\(=30\) (see seq for details). As raised by Thorpe et al. (2017) , this is one of the advantages of using the NE as the multispecies extension of the Maximum Sustainable Yield concept.

To compute the interaction matrix a second order central difference quotient is used to approximate derivatives. F.increase is employed during this calculation as a step-size set by default to \(0.1\) to avoid truncation and/or rounding errors (Pope et al. 2019) .

The `LV` method is set by default given its performance advantage (Del Santo O'Neill et al. 2023) over the `round-robin` method and is based on the protocol devised by (Farcas and Rossberg 2016) . For each species \(i\) in turn, round-robin iteratively maximises the yield by adjusting the harvesting rates whereas LV does the same simultaneously for all species per iteration.

References

Matsuda H, Abrams PA (2006). “Maximal yields from multispecies fisheries systems: rules for systems with multiple trophic levels.” Ecological Applications, 16(1), 225--237.

Worm B, Hilborn R, Baum JK, Branch TA, Collie JS, Costello C, Fogarty MJ, Fulton EA, Hutchings JA, Jennings S, Jensen OP, Lotze HK, Mace PM, McClanahan TR, Minto C, Palumbi SR, Parma AM, Ricard D, Rosenberg AA, Watson R, Zeller D (2009). “Rebuilding global fisheries.” Science, 325(5940), 578--585. doi:10.1126/science.1173146 .

Thorpe RB, Jennings S, Dolder PJ (2017). “Risks and benefits of catching pretty good yield in multispecies mixed fisheries.” ICES Journal of Marine Science, 74(8), 2097--2106. doi:10.1093/icesjms/fsx062 .

Pope JG, Bartolino V, Kulatska N, Bauer B, Horbowy J, Ribeiro JPC, Sturludottir E, Thorpe R (2019). “Comparing the steady state results of a range of multispecies models between and across geographical areas by the use of the jacobian matrix of yield on fishing mortality rate.” Fisheries Research, 209, 259--270. doi:10.1016/j.fishres.2018.08.011 .

Farcas A, Rossberg AG (2016). “Maximum sustainable yield from interacting fish stocks in an uncertain world: two policy choices and underlying trade-offs.” ICES J Mar Sci, 73(10), 2499--2508. doi:10.1093/icesjms/fsw113 .

Lucey SM, Gaichas SK, Aydin KY (2020). “Conducting reproducible ecosystem modeling using the open source mass balance model Rpath.” Ecological Modelling, 427, 109057. doi:10.1016/j.ecolmodel.2020.109057 .